Django_API_Advanced/requirements.txt
django==4.0.4
djangorestframework==3.13.1
pillow==9.1.1